// ComparaFrontend/src/app/models/producto-financiero.ts
export class ProductoFinanciero{
  consultaEntidadProductoId: number = 0;
  entidadProductoId: number = 0;
  pocentajeTea: number = 0;
  porcentajeComision: number= 0;
  porcentajeSeguro: number= 0;
  plazoMinimoMes: number= 0;
  plazoMaximoMes: number= 0;
  descripcion: string = '';
  mensajeOpcional: string= '';
  detalleProducto: string= '';
  resumenProducto: string= '';
  plazoMinimoDia: number=0;
  plazoMaximoDia: number=0;
  tieneFondoSeguro: boolean=false;
  nombreEntidadFinanciera: string= '';
  imagenEntidadFinanciera: string= '';
  tipoInstitucionId: number=0;
  departamentoId: number=0;
  codigoProductoFinanciero: string= '';
  nombreProductoFinanciero: string= '';
  descripcionProductoFinanciero: string= '';
  detalleEntidadProductoId: number=0;
  tipoMonedaId: number=0;
  tipoMoneda: string= '';
  ingresoPermitido: number=0;
  montoMinimoAceptable: number=0;
  montoMaximoAceptable: number=0;
  montoMinimoDeposito: number=0;
  montoMaximoDeposito: number=0;
}